Split ImageManagerTest
